+++ plugins/postcss-global-data/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS global-data
 
+### Unreleased (patch)
+
+- Fixed: file watching with PostCSS CLI
+
 ### 1.0.1 (February 9, 2023)
 
 - Fixed: set the source file on injected data.